+++ Changelog.md
@@ -13,6 +13,7 @@
 - Fix: Blocks projected by a Projector are reset to their default state when opening the Projector's screen
 - Fix: The owner of the topmost block of the cage spawned by a Cage Trap is not set correctly
 - Fix: Crash when interacting with certain blocks belonging to a different player
+- Fix: Certain blocks can be mined by any player, regardless of who owns the block
 
 --------------------------Changelog for v1.10 of SecurityCraft--------------------------